move and clean upload-job functions
