+++ apps/api/src/hyp3_api/util.py
@@ -2,11 +2,13 @@ import binascii
 import json
 from base64 import b64decode, b64encode
 from pathlib import Path
+from tempfile import TemporaryDirectory
 from typing import Any
 from urllib.parse import parse_qsl, urlencode, urlparse, urlunparse
 
 import boto3
 from boto3.s3.transfer import TransferConfig
+from werkzeug.datastructures import FileStorage
 
 
 S3_CLIENT = boto3.client('s3')
@@ -60,10 +62,21 @@ def get_current_account_arn() -> str:
     return sts.get_caller_identity()['Account']
 
 
-def upload_file_to_s3(
+def _upload_file_to_s3(
     path_to_file: Path, content_type: str, bucket: str, prefix: str = '', chunk_size: int = 8_388_608
 ) -> None:
     key = str(Path(prefix) / path_to_file.name)
     extra_args = {'ContentType': content_type}
     config = TransferConfig(multipart_threshold=chunk_size, multipart_chunksize=chunk_size)
     S3_CLIENT.upload_file(str(path_to_file), bucket, key, extra_args, Config=config)
+
+
+def save_and_upload_to_s3(file_obj: FileStorage, bucket: str, bucket_prefix: str) -> str:
+    filename = file_obj.filename
+    assert filename
+    with TemporaryDirectory() as temp_dir:
+        filepath = Path(temp_dir) / filename
+        file_obj.save(filepath)
+        _upload_file_to_s3(filepath, file_obj.mimetype, bucket, bucket_prefix)
+    filename = f'{bucket}/{bucket_prefix}/{filename}'
+    return filename
